// src/types/restaurant.d.ts
export type Location = {
  streetName: string;
  streetNumber: string;
  postalCode: string;
  city: string;
  country: string;
};
// ...
export type ColophonCorporation = ColophonCommons & {
  //Kapitalgesellschaft (z.B. GmbH, AG)
  // Name und Anschrift der Gesellschaft (Restaurant GmbH)
  // Vertretungsberechtigte Personen (Geschäftsführer, Vorstand)
  // Umsatzsteuer-Identifikationsnummer (falls vorhanden)
  // Handelsregister, Registergericht und Registernummer
  // Angabe des Kapitals (bei GmbH und AG)
  // Klarer Hinweis auf die Verantwortlichkeit (z.B. "Verantwortlich im Sinne des § 55 Abs. 2 RStV")
  businessType: 'Corporation';
  representativeNames: string[];
  vatNumber?: string;
  registerChamber: string;
  registerNumber: string;
};
// ...
type ColophonCommons = {
  companyName: string;
  location: Location;
  telephone: string;
  fax: string;
  email: string;
  contentResponsible: string;
};